requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.121.0
h11==0.16.0
httptools==0.7.1
idna==3.11
joblib==1.5.2
numpy==2.3.4
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
scikit-learn==1.7.2
scipy==1.16.3
sniffio==1.3.1
starlette==0.49.3
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
